# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SocialNetworkApi",
    "description": "Social Network Api for sharing information."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    {
      "name": "User",
      "description": "Endpoints"
    },
    {
      "name": "Profile",
      "description": "Endpoints"
    },
    {
      "name": "Post",
      "description": "Endpoints"
    },
    {
      "name": "PostCategory",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/posts/test": {
      "get": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to Tests post route",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/posts/": {
      "get": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to get all posts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to create post",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to Get post by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to delete post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/posts/like/{id}": {
      "post": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to like post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/posts/comment/{id}": {
      "post": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to add comment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "avatar": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/posts/comment/{id}/{comment_id}": {
      "post": {
        "tags": [
          "Post"
        ],
        "description": "Endpoint to Tremove comment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/users/test": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to test inuser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/current": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get current logged in user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/users/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to register",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "password2": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/users/Users": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/profile/test": {
      "get": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to test profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/profile/": {
      "get": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Profile"
        ],
        "description": "create or edit user profile",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "bio": {
                  "example": "any"
                },
                "githubusername": {
                  "example": "any"
                },
                "youtube": {
                  "example": "any"
                },
                "facebook": {
                  "example": "any"
                },
                "twitter": {
                  "example": "any"
                },
                "instagram": {
                  "example": "any"
                },
                "linkedin": {
                  "example": "any"
                },
                "skills": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/profile/all": {
      "get": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to Get all profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/profile/user/{user_id}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to Get profile by user ID",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to Delete user and profile",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/profile/experience": {
      "post": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to add experience to profile",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "company": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "from": {
                  "example": "any"
                },
                "to": {
                  "example": "any"
                },
                "current": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/profile/education": {
      "post": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to add education to profile",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "school": {
                  "example": "any"
                },
                "degree": {
                  "example": "any"
                },
                "fieldofstudy": {
                  "example": "any"
                },
                "from": {
                  "example": "any"
                },
                "to": {
                  "example": "any"
                },
                "current": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/profile/experience/{exp_id}": {
      "delete": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to delete experience",
        "parameters": [
          {
            "name": "exp_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/profile/education/{id}": {
      "delete": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to delete education",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/postcategorys/test": {
      "get": {
        "tags": [
          "PostCategory"
        ],
        "description": "Endpoint to Tests postcategory route",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/postcategorys/": {
      "get": {
        "tags": [
          "PostCategory"
        ],
        "description": "Endpoint to Get All postcategory",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "PostCategory"
        ],
        "description": "Endpoint to create postcategory ",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/postcategorys/{id}": {
      "get": {
        "tags": [
          "PostCategory"
        ],
        "description": "Endpoint to Get postscategory by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  }
}